refact, move runner to utils
